alter file blade

+++ public/js/article.js
@@ -1,11 +1,18 @@
-const $ = document.querySelector.bind(document)
-const $$ = document.querySelectorAll.bind(document)
-import { mess } from './main.js'
+// const $ = document.querySelector.bind(document)
+// const $$ = document.querySelectorAll.bind(document)
+// import { mess } from './main.js'
 // can sua 
-const user_id = 2
+const article_id = $('.box-post').getAttribute('data-id-article')
+const comment = $('.comment-box')
 
 function start() {
     getPost()
+    getComments()
+    $('.sent-comment').addEventListener('click', () => {
+        console.log($('.sent-comment'))
+        createComment()
+    })
+    handleTotalLike()
 }
 
 function getPost() {
@@ -15,9 +22,8 @@ function getPost() {
         .then(data => {
             const author_id = data.author_id;
             let status = data.status_article;
-            // $('#categories').innerHTML = data.data_article[0].data.text
             let htmls = `
-                <h1 id="categories" data-cate-id="${data.data_user[0].categorie_id}">${data.data_article[0].data.text}</h1>
+                <h1 id="categories" data-cate-id="${data.categorie_id}">${data.data_article[0].data.text}</h1>
                 <div class="info-author">
                         <div class="author">
                             <img src="${data.data_user[0].url}"
@@ -49,7 +55,7 @@ function getPost() {
             }).join('')
             //  them btn chinh sua, xoa va dang
             htmls += `<div class="list-btn">
-            </div>` 
+            </div>`
 
 
             $('.box-list-item').innerHTML = htmls
@@ -57,7 +63,6 @@ function getPost() {
             return [status, author_id, data];
         })
         .then(data => {
-            console.log(data[2].id)
             if (data[0] === 1) {
                 $('#status').innerHTML = 'Bản nháp'
                 if (data[1] === user_id) {
@@ -79,18 +84,20 @@ function getPost() {
             }
             else {
                 $('#status').style.display = 'none'
-                if (data[1] === user_id) { 
+                if (data[1] === user_id) {
                     $('.list-btn').innerHTML = `
                         <a href="http://127.0.0.1:8000/index/editer/write_article/${data[2].id}"><button class="btn-item">Chỉnh sửa</button></a>
                         <button class="btn-item delete-article">Xóa</button>
                     `
                 }
             }
-            $('.delete-article').onclick = () => { 
-                deleteArticle(data[2].id)
+            if (user_id) {
+                $('.delete-article').onclick = () => {
+                    deleteArticle(data[2].id)
+                }
             }
             if ($('.post-article'))
-                $('.post-article').onclick = () => { 
+                $('.post-article').onclick = () => {
                     postArticle(data[2].id)
                 }
         })
@@ -98,7 +105,7 @@ function getPost() {
 
 function deleteArticle(id) {
     let result = confirm('Bạn có chắc muốn xóa vĩnh viễn bài viết')
-    if (result) { 
+    if (result) {
         const options = {
             method: 'DELETE',
             headers: {
@@ -137,6 +144,144 @@ function postArticle(id) {
                 alert('Đăng bài không thành công!' + ' \nLỗi:' + err)
             })
     }
+}
+
+function getComments() {
+    fetch(`http://127.0.0.1:8000/api/comment/${article_id}`)
+        .then(response => response.json())
+        .then(comments => {
+            let i
+            const htmls = comments.map((cm, index) => {
+                if (cm.user) {
+                    const hours = Math.floor(Math.abs(new Date().getTime() - (new Date(cm.created_at))) / (3600 * 1000))
+                    let day;
+                    let dateTime = hours + ' Giờ trước';
+                    if (hours >= 24) {
+                        day = Math.floor(hours / 24)
+                        dateTime = day + ' Ngày trước';
+                    }
+                    return `
+                        <div class="cm-item">
+                            <img src="${cm.user[0].url}" alt="">
+                            <div class="cm-item-info-feeling">
+                                <div class="cm-item-info">
+                                    <h5 class="name">${cm.user[0].name}</h5>
+                                    <p>${cm.content}</p>
+                                </div>
+                                <ul class="cm-item-feeling">
+                                    <li>Thích</li>
+                                    <li>Phản hồi</li>
+                                    <li>${dateTime}}</li>
+                                    <li>
+
+                                    </li>
+                                </ul>
+                            </div>
+                        </div>
+                    `
+                }
+                i = index
+            })
+
+            $('.info-post-box p:nth-child(1)').innerHTML = `${i} Bình luận`
+            comment.innerHTML = htmls.join('')
+            return comment
+        })
+        .then((comment) => {
+            const cmItem = $$('.comment-box .cm-item')
+            cmItem.forEach((element, i) => {
+                if (i > 4)
+                    element.style.display = 'none'
+            });
+            const showAll = document.createElement("p")
+            showAll.innerHTML = 'Xem tất cả bình luận...'
+            showAll.classList.add('show-all-cm')
+            showAll.style.cursor = 'pointer'
+            comment.appendChild(showAll)
+
+            showAll.onclick = () => {
+                cmItem.forEach((element) => {
+                    element.style.display = 'flex'
+                });
+                comment.removeChild(showAll)
+            }
+            return comment
+        })
+        .then((cm) => {
+            addImgComment()
+        })
+}
+
+function createComment() {
+    const content = $('input[name="content-comment')
+    const data = {
+        user_id: user_id,
+        article_id: article_id,
+        content: content.value
+    }
+    const options = {
+        method: 'POST',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify(data)
+    }
+    if (String(content.value).trim()) {
+        fetch('http://127.0.0.1:8000/api/comment', options)
+            .then(response => response.json())
+            .then(data => {
+                let addNewComment
+                Array(data).forEach((cm) => {
+                    const hours = Math.floor(Math.abs(new Date().getTime() - (new Date(cm.created_at))) / (3600 * 1000))
+                    let day;
+                    let dateTime = hours + ' Giờ trước';
+                    if (hours >= 24) {
+                        day = Math.floor(hours / 24)
+                        dateTime = day + ' Ngày trước';
+                    }
+                    addNewComment = `
+                        <div class="cm-item">
+                            <img src="${cm.user[0].url}" alt="">
+                            <div class="cm-item-info-feeling">
+                                <div class="cm-item-info">
+                                    <h5 class="name">${cm.user[0].name}</h5>
+                                    <p>${cm.content}</p>
+                                </div>
+                                <ul class="cm-item-feeling">
+                                    <li>Thích</li>
+                                    <li>Phản hồi</li>
+                                    <li>${dateTime}</li>
+                                </ul>
+                            </div>
+                        </div>
+                `
+                })
+                comment.insertAdjacentHTML('afterbegin', addNewComment);
+            })
+            .then(() => {
+                content.value = ''
+            })
+    }
+}
+
+async function addImgComment() {
+    fetch(`http://127.0.0.1:8000/api/user/${login_id}`)
+        .then(response => response.json())
+        .then(data => {
+            console.log(data)
+            $('.write-comment img').src = data[0].url;
+
+        })
+}
+
+function handleTotalLike() {
+    fetch(`http://127.0.0.1:8000/api/like/${article_id}/${user_id}`)
+        .then(response => response.json())
+        .then(data => { 
+            console.log(JSON.parse(data.action))
+            $('.info-post p:nth-child(1)').innerHTML = data.total + ' Lượt thích'
+            if (JSON.parse(data.action)) {
+                $('.feeling-box').classList.add('active-emotion')
+            }
+        })
 }
 
 start()
